fix code structure
